firstpage_design

+++ app/src/main/java/com/jjthedev/pilly/Pill.java
@@ -4,8 +4,9 @@ import java.util.List;
 
 public class Pill {
     public String name;
-    public float dosage;
+    public int dosage;
     public int count;
 
     public List<Integer> timings;
+
 }
